+++ src/si5351.h
@@ -282,7 +282,7 @@ public:
   Si5351(uint8_t i2c_addr = SI5351_BUS_BASE_ADDR);
 	void init(uint8_t, uint32_t, int32_t);
 	void reset(void);
-	uint64_t set_freq(uint64_t, enum si5351_clock);
+	uint8_t set_freq(uint64_t, enum si5351_clock);
 	uint8_t set_freq_manual(uint64_t, uint64_t, enum si5351_clock);
 	void set_pll(uint64_t, enum si5351_pll);
 	void set_ms(enum si5351_clock, struct Si5351RegSet, uint8_t, uint8_t, uint8_t);
